+++ libhearingdownloader.py
@@ -28,8 +28,8 @@ import os
 # libhearingdownloader - A useful library for the downloader scripts
 ###
 
-#downloaderVersion = "Pre-release"
-downloaderVersion = "v2025.06.10"
+downloaderVersion = "Pre-release"
+#downloaderVersion = "v2025.06.10"
 updaterRetries = 3
 # Read configuration file for toggles with default True
 config = configparser.ConfigParser()
@@ -228,7 +228,6 @@ def downloadFile(url, saveLocation, downloadDescription):
 
         streamRetries -= 1
     if (streamRetries == 0):
-        print("\n" + Fore.RED + "Error" + Style.RESET_ALL + ": Can not get proper file data stream")
+        print("\n" + Fore.RED + "Error" + Style.RESET_ALL + ": Can not get proper file data stream from " + Fore.GREEN + url + Style.RESET_ALL + "\nYou can try to open this URI manually with web browsers. It might work.")
         exit(1)
-